+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-v1.3.1 (2022.08.xx)
+v1.3.1 (2022.08.26)
 -------------------
 - Add variants to media
 - Add alt_text to media